+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/NftRedPacketHistoryItem.tsx
@@ -5,7 +5,8 @@ import { makeStyles, ActionButton } from '@masknet/theme'
 import { NetworkPluginID } from '@masknet/shared-base'
 import { isSameAddress, type NonFungibleCollection } from '@masknet/web3-shared-base'
 import { useChainContext, useNetworkDescriptor } from '@masknet/web3-hooks-base'
-import type { ChainId, SchemaType, NftRedPacketJSONPayload } from '@masknet/web3-shared-evm'
+import type { ChainId, SchemaType } from '@masknet/web3-shared-evm'
+import { type NftRedPacketJSONPayload } from '@masknet/web3-providers/types'
 import { Box, ListItem, Typography } from '@mui/material'
 import { dateTimeFormat } from '../../ITO/assets/formatDate.js'
 import { Translate, useI18N } from '../locales/index.js'
