+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/NftRedPacketHistoryList.tsx
@@ -4,7 +4,8 @@ import { useSharedI18N } from '@masknet/shared'
 import { EMPTY_LIST, NetworkPluginID } from '@masknet/shared-base'
 import { useChainContext, useNonFungibleCollections } from '@masknet/web3-hooks-base'
 import type { NonFungibleCollection } from '@masknet/web3-shared-base'
-import type { ChainId, SchemaType, NftRedPacketJSONPayload } from '@masknet/web3-shared-evm'
+import type { ChainId, SchemaType } from '@masknet/web3-shared-evm'
+import { type NftRedPacketJSONPayload } from '@masknet/web3-providers/types'
 import { List, Popper, Typography, Box } from '@mui/material'
 import { useNftRedPacketHistory } from './hooks/useNftRedPacketHistory.js'
 import { NftRedPacketHistoryItem } from './NftRedPacketHistoryItem.js'
