+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/RedPacketPast.tsx
@@ -3,7 +3,8 @@ import { TabPanel } from '@mui/lab'
 import { makeStyles } from '@masknet/theme'
 import { Box } from '@mui/material'
 import type { NonFungibleCollection } from '@masknet/web3-shared-base'
-import type { ChainId, SchemaType, RedPacketJSONPayload, NftRedPacketJSONPayload } from '@masknet/web3-shared-evm'
+import type { ChainId, SchemaType } from '@masknet/web3-shared-evm'
+import { type RedPacketJSONPayload, type NftRedPacketJSONPayload } from '@masknet/web3-providers/types'
 import { RedPacketHistoryList } from './RedPacketHistoryList.js'
 import { NftRedPacketHistoryList } from './NftRedPacketHistoryList.js'
 import { RedPacketNftMetaKey } from '../constants.js'
