+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/hooks/useNftAvailabilityComputed.ts
@@ -3,7 +3,8 @@ import { compact } from 'lodash-es'
 import { useChainContext } from '@masknet/web3-hooks-base'
 import { isSameAddress } from '@masknet/web3-shared-base'
 import type { NetworkPluginID } from '@masknet/shared-base'
-import { ChainId, chainResolver, type NftRedPacketJSONPayload, RedPacketStatus } from '@masknet/web3-shared-evm'
+import { ChainId, chainResolver } from '@masknet/web3-shared-evm'
+import { type NftRedPacketJSONPayload, RedPacketStatus } from '@masknet/web3-providers/types'
 import { useAvailabilityNftRedPacket } from './useAvailabilityNftRedPacket.js'
 
 /**
