+++ packages/mask/src/plugins/RedPacket/SNSAdaptor/utils/reduceUselessPayloadInfo.ts
@@ -1,5 +1,6 @@
 import type { FungibleToken } from '@masknet/web3-shared-base'
-import type { ChainId, SchemaType, RedPacketJSONPayload } from '@masknet/web3-shared-evm'
+import type { ChainId, SchemaType } from '@masknet/web3-shared-evm'
+import { type RedPacketJSONPayload } from '@masknet/web3-providers/types'
 import { pick, omit } from 'lodash-es'
 
 export function reduceUselessPayloadInfo(payload: RedPacketJSONPayload): RedPacketJSONPayload {
