+++ packages/mask/src/plugins/RedPacket/Worker/databaseForNft.ts
@@ -1,5 +1,5 @@
 import { omit } from 'lodash-es'
-import type { RedPacketNftRecordInDatabase, RedPacketNftRecord } from '@masknet/web3-shared-evm'
+import type { RedPacketNftRecordInDatabase, RedPacketNftRecord } from '@masknet/web3-providers/types'
 import { RedPacketDatabase } from './database.js'
 
 export async function getRedPacketNft(id: string) {
