+++ packages/mask/src/plugins/RedPacket/Worker/servicesForNft.ts
@@ -1,4 +1,4 @@
-import type { RedPacketNftRecord, RedPacketNftRecordInDatabase } from '@masknet/web3-shared-evm'
+import type { RedPacketNftRecord, RedPacketNftRecordInDatabase } from '@masknet/web3-providers/types'
 import * as database from './databaseForNft.js'
 
 export async function addRedPacketNft(record: RedPacketNftRecord) {
