+++ packages/web3-providers/src/entry-types.ts
@@ -46,6 +46,7 @@ export * from './Zerion/types.js'
 export * from './RSS3/types.js'
 export * from './GoPlusLabs/types.js'
 export * from './Web3/Router/types/index.js'
+export * from './RedPacket/types.js'
 
 export type { ConnectionOptions_Base } from './Web3/Base/apis/ConnectionOptionsAPI.js'
 export type { HubOptions_Base } from './Web3/Base/apis/HubOptionsAPI.js'
