+++ packages/web3-shared/evm/src/index.ts
@@ -1,5 +1,4 @@
 export * from './types/index.js'
-export * from './types/RedPacket.js'
 export * from './constants/index.js'
 export * from './libs/index.js'
 export * from './helpers/address.js'
